Class Debloat: Interface Extraction
I've extracted the product csv handling functionality and the HashPassword functionality into their own interfaces.

(Also yes I dontnet cleaned)
