+++ .gitignore
@@ -1,3 +1,5 @@
 obj/
 /.vs/IMS/copilot-chat
 /.vs/IMS/CopilotIndices
+/.vs/IMS/CopilotIndices
+/.vs/IMS/CopilotIndices
