+++ Interfaces/IHashPasswords.cs
@@ -1,0 +1,22 @@
+using System.Security.Cryptography;
+using System.Text;
+
+namespace IMS.Interfaces {
+    public interface IHashPasswords
+    {
+
+        static string HashPassword(string password)
+        {
+            using (var sha256 = SHA256.Create())
+            {
+                var bytes = sha256.ComputeHash(Encoding.UTF8.GetBytes(password));
+                var builder = new StringBuilder();
+                for (int i = 0; i < bytes.Length; i++)
+                {
+                    builder.Append(bytes[i].ToString("x2"));
+                }
+                return builder.ToString();
+            }
+        }
+    }
+}
