+++ Pages/Error.cshtml.cs
@@ -1,4 +1,5 @@
 using System.Diagnostics;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
@@ -6,6 +7,7 @@ namespace IMS.Pages;
 
 [ResponseCache(Duration = 0,Location = ResponseCacheLocation.None,NoStore = true)]
 [IgnoreAntiforgeryToken]
+[AllowAnonymous]
 public class ErrorModel:PageModel {
   public string? RequestId { get; set; }
