+++ Pages/Privacy.cshtml.cs
@@ -1,8 +1,10 @@
 ﻿using IMS.Attributes;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
 namespace IMS.Pages;
+[AllowAnonymous]
 public class PrivacyModel:PageModel {
   private readonly ILogger<PrivacyModel> _logger;
