+++ Pages/Products.cshtml.cs
@@ -13,42 +13,11 @@ using CsvHelper.Configuration;
 using System.Text.RegularExpressions;
 using static Microsoft.EntityFrameworkCore.DbLoggerCategory;
 using Microsoft.IdentityModel.Tokens;
+using IMS.Interfaces;
 
 namespace IMS.Pages {
-    public class ProductsModel : PageModel {
+    public class ProductsModel : PageModel, IHandleProductCsv {
         private readonly AppDbContext _appDbContext;
-
-        private class ProductMap : ClassMap<Product>{
-            public ProductMap(bool export = false){
-                Map(m => m.Name).Default(string.Empty);
-                Map(m => m.Description).Default(string.Empty);
-                Map(m => m.Price).Default(0.00m);
-                Map(m => m.Quantity).Default(0);
-                Map(m => m.ReorderLevel).Default(0);
-                Map(m => m.SKU).Default(string.Empty);
-                Map(m => m.Category).Default(string.Empty);
-                Map(m => m.Location).Default(string.Empty);
-                if (export)
-                {
-                    //convert binary to base64 when used for CSV export
-                    Map(m => m.Image).Convert(args =>
-                        args.Value.Image != null && args.Value.Image.Length > 0
-                            ? Convert.ToBase64String(args.Value.Image)
-                            : "");
-                }
-                else
-                {
-                    //convert uploaded Base64 to byte array when used for CSV import
-                    Map(m => m.Image).Convert(args =>
-                    {
-                        var base64 = args.Row.GetField("Image");
-                        return string.IsNullOrEmpty(base64)
-                            ? Array.Empty<byte>()
-                            : Convert.FromBase64String(base64);
-                    });
-                }
-            }
-        }
 
         public ProductsModel(AppDbContext appDbContext) {
             _appDbContext = appDbContext;
@@ -125,7 +94,7 @@ namespace IMS.Pages {
                 await PopulateProductsList();
                 return Page();
             }
-
+           
             var product = await _appDbContext.Products.FindAsync(Input.Id);
 
             if (product == null) {
@@ -374,164 +343,28 @@ namespace IMS.Pages {
                 return Page();
             }
 
-
-            using (var reader = new StreamReader(file.OpenReadStream())) {
-                
-                CsvReader csv = new(reader, new CsvHelper.Configuration.CsvConfiguration(CultureInfo.InvariantCulture)
-                {
-                    Delimiter = ",",
-                    PrepareHeaderForMatch = args => args.Header.Trim(),
-                    HasHeaderRecord = true,
-                    MissingFieldFound = null, // Ignore missing fields
-                    HeaderValidated = null, // Ignore header validation
-                });
-
-                csv.Read(); //read the first row
-                csv.ReadHeader(); //read the first row as the header row
-
-                // Validate the header
-                string[]? headerRecord = csv.HeaderRecord;
-                if(!ValidateCsvHeader(headerRecord))
-                {
-                    ViewData["Errors"] = ModelState.Values.SelectMany(v => v.Errors).Select(e => e.ErrorMessage).ToList();
-                    await PopulateProductsList();
-                    return Page();
-                }
-
-                
-                csv.Context.RegisterClassMap(new ProductMap()); //handles converting empty csv fields to default values. 
-
-                //attempt to read the records as Product objects
-                List<Product> records;
-                try
-                {
-                    records = csv.GetRecords<Product>().ToList();
-                }
-                catch (Exception ex)
-                {
-                    ModelState.AddModelError(string.Empty, $"CSV parsing error: {ex.Message}");
-                    ViewData["Errors"] = ModelState.Values.SelectMany(v => v.Errors).Select(e => e.ErrorMessage).ToList();
-                    await PopulateProductsList();
-                    return Page();
-                }
-
-                // Validate records before saving, if any records are invalid, return to the page with errors
-                if (!ValidateRecords(records))
-                {
-                    ViewData["Errors"] = ModelState.Values.SelectMany(v => v.Errors).Select(e => e.ErrorMessage).ToList();
-                    await PopulateProductsList();
-                    return Page();
-                }
-                _appDbContext.Products.AddRange(records);
+            try
+            {
+               List<Product> products = IHandleProductCsv.ConvertFromCsv(file);
+                _appDbContext.Products.AddRange(products);
                 await _appDbContext.SaveChangesAsync();
+
+                return RedirectToPage("/Products");
             }
-
-            return RedirectToPage("/Products");
-        }
-
-
-
-
-        private bool ValidateRecords(List<Product> records) {
-            bool isValid = true;
+            catch (ProductCsvException ex)
+            {
+               //foreach (string err in ex.Errors) ModelState.AddModelError(string.Empty, err);
+                ViewData["Errors"] = ex.Errors;
+                await PopulateProductsList();
+                return Page();
+            }
             
-            for(int i = 0; i < records.Count; i++)
-            //foreach (var record in records)
-            {
-                Product record = records[i];
-                int index = i + 1; // 1-based index for error messages
-                if (string.IsNullOrEmpty(record.Name)){
-                    ModelState.AddModelError(string.Empty, $"Item {index}: Name is a required field and cannot be empty.");
-                    isValid = false;
-                }
-
-                if (record.Price < 0)
-                {
-                    ModelState.AddModelError(string.Empty, $"Item {index}: An item's Price cannot be negative.");
-                    isValid = false;
-                }
-                if (record.Quantity < 0)
-                {
-                    Console.WriteLine($"Validation failed for Quantity at index: {index}");
-                    ModelState.AddModelError(string.Empty, $"Item {index}: An item's Quantity cannot be negative.");
-                    isValid = false;
-                }
-                if (record.ReorderLevel < 0)
-                {
-                    ModelState.AddModelError(string.Empty, $"Item {index}: An item's Reorder Level cannot be negative.");
-                    isValid = false;
-                }
-
-                if (!string.IsNullOrEmpty(record.SKU))
-                {
-                    if (!Regex.IsMatch(record.SKU, @"^[A-Z0-9-]*$") ||
-                        !Regex.IsMatch(record.SKU, @"^\S+(-\S+)+$") ||
-                        !Regex.IsMatch(record.SKU, @"^\S{1,5}(-\S*)*$") ||
-                        !Regex.IsMatch(record.SKU, @"^\S+(-\S{1,10})+$") ||
-                        !Regex.IsMatch(record.SKU, @"^\S+(-\S+){1,3}$"))
-                    {
-                        ModelState.AddModelError(string.Empty, $"Item {index}: Incorrectly formatted SKU.");
-                        isValid = false;
-                    }
-                }
-            }
-            return isValid;    
         }
-
-        private bool ValidateCsvHeader(string[]? headerRecord)
-        {
-            bool isValid = true;
-            if (headerRecord == null)
-            {
-                ModelState.AddModelError(string.Empty, "No header record found.");
-                return false;
-            }
-            var allowedHeaders = new List<string> {
-                "Name", "Description", "Price", "Quantity", "ReorderLevel", "SKU", "Category", "Location", "Image"
-            };
-
-            //Name header is required
-            if (!headerRecord.Contains("Name", StringComparer.OrdinalIgnoreCase))
-            {
-                ModelState.AddModelError(string.Empty, "The 'Name' header is required but missing.");
-                isValid = false;
-            }
-
-            // all other headers besides Name are optional, but if they are present, they must match the allowedHeaders list
-            var invalidHeaders = headerRecord
-                .Where(header => !allowedHeaders.Contains(header, StringComparer.OrdinalIgnoreCase))
-                   .ToList();
-
-            if (invalidHeaders.Any())
-            {
-                ModelState.AddModelError(string.Empty, $"Invalid headers found: {
-                    string.Join(", ", invalidHeaders.Select(header => $"\"{header}\""))
-                    }.");
-                isValid = false;
-            }
-
-
-            var duplicateHeaders = headerRecord
-                .GroupBy(header => header, StringComparer.OrdinalIgnoreCase)
-                .Where(group => group.Count() > 1)
-                .Select(group => group.Key);
-
-            if (duplicateHeaders.Any())
-            { 
-                ModelState.AddModelError(string.Empty, $"Duplicate headers found: {
-                    string.Join(", ", duplicateHeaders.Select(header => $"\"{header}\""))
-                    }.");
-                isValid = false; // Duplicate headers found
-            }
-
-            return isValid;
-        }
-
 
         public async Task<IActionResult> OnPostDownloadCSVAsync(List<int> productIds)
         {
             //previous product list is entered as hidden form, this will allow downloading of narrowed search results
-            Products = await _appDbContext.Products
+            List<Product> Products = await _appDbContext.Products
                 .Where(p => productIds.Contains(p.Id))
                 .ToListAsync();
 
@@ -541,19 +374,7 @@ namespace IMS.Pages {
                 await PopulateProductsList();
             }
 
-
-            using (StringWriter writer = new StringWriter())
-            using (CsvWriter csv = new CsvWriter(writer, CultureInfo.InvariantCulture))
-            {
-                csv.Context.RegisterClassMap(new ProductMap(true));
-                csv.WriteRecords(Products);
-
-                string fileName = $"Products_{DateTime.Now:yyyyMMdd_HHmmss}.csv";
-                string contentType = "text/csv";
-                string fileContent = writer.ToString();
-                return File(new System.Text.UTF8Encoding().GetBytes(fileContent), contentType, fileName);
-            }
-
+            return IHandleProductCsv.ConvertToCsv(Products);
         }
     }
 }
